Add custom offchain extention to test-helper
